requirements.txt
# Core dependencies
pandas>=2.2.0
numpy>=1.26.0
networkx>=3.2.1
scikit-learn>=1.4.0
requests>=2.31.0

# Geospatial and mapping
geopandas>=0.14.1
folium>=0.15.1
geopy>=2.4.1
shapely>=2.0.2

# Web framework and visualization
streamlit>=1.29.0
plotly>=5.17.0
dash>=2.16.1
bokeh>=3.3.2

# Data processing
gtfs-realtime-bindings>=1.0.0
gtfs-kit>=6.1.0

# API and web services
fastapi>=0.108.0
uvicorn>=0.25.0
httpx>=0.26.0

# Database and storage
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
redis>=5.0.1

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.2
click>=8.1.7

# Testing and development
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0

# Logging and monitoring
structlog>=23.2.0
sentry-sdk>=1.39.2

# Machine learning
joblib>=1.3.2
xgboost>=2.0.3
lightgbm>=4.2.0

# Deployment
docker>=7.0.0
gunicorn>=21.2.0